+++ java-service-processor/src/test/resources/definition/LostId.java
@@ -1,0 +1,9 @@
+package definition;
+
+import nbbrd.service.ServiceId;
+
+interface LostId {
+
+    @ServiceId
+    String getName();
+}