+++ java-service-processor/src/test/resources/definition/NoArgId.java
@@ -1,0 +1,11 @@
+package definition;
+
+import nbbrd.service.ServiceDefinition;
+import nbbrd.service.ServiceId;
+
+@ServiceDefinition
+interface NoArgId {
+
+    @ServiceId
+    String getName(String arg);
+}