+++ java-service-processor/src/test/resources/definition/NonStringId.java
@@ -1,0 +1,11 @@
+package definition;
+
+import nbbrd.service.ServiceDefinition;
+import nbbrd.service.ServiceId;
+
+@ServiceDefinition
+interface NonStringId {
+
+    @ServiceId
+    Object getName();
+}